Recognizing the social fabric (*.readme)
Social networks already existed back then, only that people met in the arena. As they did so, they lowered their thumbs to acknowledge mercy and held up the same thumb when they asked for more. Today it is the same and the associated raging applause has become more silent, while the protest is quietly drowned in this noise.
Unlike the majority, I don't think we can rule the fiery Internet. As long as it has a breeding ground, it cannot be extinguished. On the other hand, everyone is fighting against containment and restrictions of any kind. We cannot have both, neither a kindled fire that cannot be controlled, nor an Internet that should be extinguished. We can cool a fire if we are willing to live with the smoldering fire. But is it right to tame the fiery Internet in this way? Reaching a common consensus can take forever. And so mankind has again conquered a productive breeding ground. No one has yet succeeded in separating the spark from the fire. Even if mankind has managed to split the atom in a controlled manner, the divided material cannot be controlled or even disposed of permanently. It is the same with the half-life (tH) of individual content or the acceptance factor (qH) of content from the Internet. Both are merely to be shared in virtual places on the Internet. Such a sparkle jumps over to everyone and all places. Accept it, but don't respect it. Neither is mere copying, not to be equated with thinking!
You, the mankind came into contact with it and saw unlimited possibilities in it. And so, the mankind take responsibility for it. But you should only win over it when you made yourselves scarce. In order to be able to hold on to it, I give mankind a confession to start with. It is not to be understood as an Internet Evangelium, because this cannot be agreed globally. But very well, it is an attack of common sense on the real-time of the captured world by placarding 100 and more propositions.